% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident}
\alias{accident}
\title{Fatality Analysis Reporting System (FARS)}
\format{
A list of tibbles
\describe{
\item{2013}{A tibble: 30,202 x 50}
\item{2014}{A tibble: 30,056 x 50}
\item{2015}{A tibble: 32,166 x 52}
...
}
}
\source{
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\usage{
accident
}
\description{
FARS is a nationwide census providing NHTSA,
Congress and the American public yearly data regarding fatal
injuries suffered in motor vehicle traffic crashes.
}
\keyword{datasets}
