% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer between 1 and 56 indicating the index of the US
state (or district) to be plotted. An error message is raised if an invalid
state index is specified.}

\item{year}{integer, or convertible to integer containing the year for which
to plot the accident dataset. An error message is created if there are no
data available for that year}
}
\value{
Plot of a selected US state with the geographical location of every
accident for a given year superimposed
}
\description{
Generate a plot of the geographical distribution of accidents
for a selected US state for a given year
}
\examples{
fars_map_state(1, 2013)

}
