% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{string, full path to file or file name in project directory}
}
\value{
(if successful) dplyr::tbl_df format containing the data
extracted form the file
}
\description{
Attempt to import data from a csv file
with the given file name. Raises exception if file does not exist
}
\examples{
fars_read(file.path("data-raw", "accident_2013.csv.bz2"))

}
