% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{list containing a list of integers, or
labels convertible to integers, representing the year labels of the
accident data files}
}
\value{
returns a list of the same dimension as 'years', containing
dplyr::tbl_df objects with year and MONTH columns of the underlying data,
or NULL if no data are found
}
\description{
Iterate over a given collection of year labels,
and for each one try to extract tables of year and month labels
associated with any accident data that can be found. In case
of a missing file for a given year, a warning message is raised
and a null table is associated with that year
}
\examples{
fars_read_years(list(2013, 2014, 2015))

}
