% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{list containing a list of integers, or
labels convertible to integers, representing the year labels of the
accident data files. Warning messages will be generated if no files are found
for a given year}
}
\value{
a dplyr::tbl_df object containing the number of records found in
the files indicated by 'years' broken down by year (columns) and by month (rows)
}
\description{
Read in dataset over a given collection of year labels
and compute a count of the number of records found by year and MONTH
}
\examples{
fars_summarize_years(list(2013, 2014, 2015))

}
