% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer, or convertible to integer containing the year
label to be assigned to accident data file name}
}
\value{
string containing the file name
}
\description{
Generate a string containing the accident data file name and
expected extension labeled by the given year
}
\examples{
make_filename(2013)

}
